Moved interfaces to "types.ts"
