--- src/Data/RawItem.ts
@@ -1,4 +1,0 @@
-export interface RawItem {
-  _id: string
-  [key: string]: unknown
-}
